backend/requirements.txt
# Web Framework dependencies
flask<3.0.0
werkzeug>=2.0.0,<3.0.0
flask-cors==3.0.10

# Core ML/DL dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
scipy>=1.7.0

# Data manipulation and analysis
pandas>=1.3.0
PyYAML>=5.4.1

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
opencv-python>=4.1.1
pillow>=9.0.0

# Development tools
ipykernel~=6.29.0
ipython<8.19.0
jupyter_client>=7.0.0
jupyter_core>=4.9.0

# Utility packages
tqdm>=4.64.0
requests>=2.27.0
pathlib>=1.0.1
filelock>=3.4.0
psutil>=5.8.0

# Optional dependencies
packaging>=21.0
typing_extensions>=4.0.0
urllib3>=1.26.0
sympy~=1.13.3
networkx~=3.2.1
ultralytics>=8.0.0